Refactor kitchensink into one single route file
- Removes test implementations from main components
- Allows to restrict access to the /kitchensink soon
