+++ app/routes.ts
@@ -33,18 +33,12 @@ export default [
       ]),
       route(":id", "./routes/verfahren/$id/_layout.tsx", [
         index("./routes/verfahren/$id/_index.tsx"),
-        route(
-          "dokument/:dokumentId",
-          "./routes/verfahren/$id/dokument/$dokumentId.tsx",
-        ),
         route("bearbeiten", "./routes/verfahren/$id/bearbeiten/_layout.tsx", [
           index("./routes/verfahren/$id/bearbeiten/_index.tsx"),
         ]),
       ]),
     ]),
     // kitchensink (component playground)
-    route("kitchensink", "./routes/kitchensink/_layout.tsx", [
-      index("./routes/kitchensink/_index.tsx"),
-    ]),
+    route("kitchensink", "./routes/kitchensink.tsx"),
   ]),
 ] satisfies RouteConfig;
