--- app/routes/kitchensink/_layout.tsx
@@ -1,12 +1,0 @@
-import { Outlet } from "react-router";
-import { MatchHandle } from "~/components/Breadcrumbs";
-
-export const handle: MatchHandle = {
-  breadcrumb: {
-    title: "Kitchensink",
-  },
-};
-
-export default function KitchensinkLayout() {
-  return <Outlet />;
-}
